refactor(core): move the object cache module under api
The decoded-object cache's lifecycle is owned by the API layer (its
readers scope the L1, its Database owns the L2), so the module moves
from crate::cache to api::cache. The crate root re-exports it
(pub use api::cache), keeping the public dr_strange_core::cache path
and every internal crate::cache:: import working unchanged.
