+++ crates/dr-strange-core/src/api/cache/mod.rs
