+++ crates/dr-strange-core/src/api/cache/store.rs
